Move /api/v1/filters to /api/v3/filters
Fixes #1545
