+++ normandy/base/tests/api/test_api.py
@@ -25,7 +25,6 @@ class TestApiRootV1(object):
             "reciperevision-list": "http://testserver/api/v1/recipe_revision/",
             "classify-client": "http://testserver/api/v1/classify_client/",
             "approvalrequest-list": "http://testserver/api/v1/approval_request/",
-            "filters": "http://testserver/api/v1/filters/",
         }
 
     def test_it_redirects_classify_client_to_app_server(self, api_client, settings):
